Move TraceHTTP to the utils package
TraceHTTP only depends on net/http and httpx, with no courier-package
types, so it belongs alongside the other HTTP helpers in utils rather
than in the root package. Callers now use utils.TraceHTTP.
