+++ utils/http.go
@@ -1,4 +1,4 @@
-package courier
+package utils
 
 import (
 	"io"
